feat: verify cri error for pod manifest injection in reconcile (#3475)
Fxies: CORE-160

Move the "CriError" check from the pods webhook to the agentEnabled
reconciler

- If there is any error, it can be seen in the UI and persist on the
instrumentation config (unlink now that the error is only seen in
instrumentor logs).
- less checks and complexity in webhook to simplify it (will be done in
followup PR)
- align with loader injection which is also verified in the reconcile

There is, however, a very edge case that is being sacrificed for this
simplicity: webhook will check if the env is already at the manifest at
the time of the webhook and will append odigos value in this case, even
if there is cri-error.

The reasons I think it ok to skip this:
- if there is any env in the manifest, it should be detected in runtime
detection and we should NOT get criError since we can tell that envs are
not coming from cri.
- We never actually seen a cri error during the 9 months since this
feature was introduced
- This case is only for when a user has CRI-error, env set in `/proc`
somehow but not in manifest, and then he made a change to include it in
the manifest. at this point the current implementation would work, but
new implementation will not instrument. I believe this scenario is not
something that we will ever encounter, and even if we do, the user is
already at an edge case and uninstrumenting and instrumenting is the
path to go to resolve this (re-calculate runtime details)