feat: record time at which agent meta has changed (#3479)
Fixes: CORE-247

This PR records another property to instrumentationconfig to denote the
time at which the agents meta hash changed.

When listing pods for a workload, and observing the presence of the
odigos `odigos.io/agents-meta-hash` label, it is useful to be able to
classify if the observed value is correct or not.

For example, a pod might not have the label, but it's expected if it
started before we enabled the agent with the updated hash. On the other
end, if the pod is labeled when it shouldn't be, but it was started at a
time when the agent injection was enabled, then it is expected that the
pod is still instrumented.

The new timestamp allows us to identify this situation correctly to
differentiate between error types in the ui later on.