chore: refactor distro params handling (#3478)
Fixes: CORE-160

Distro params are a way to propagate parameters from the runtime
detection into the webhook, to be used during the injection of the
distro in the webhook.

It allows use to:
- verify the value from runtime detection is available when required
- the value is in the right format (for example - a version number)
- process the value to make it useful for the webhook (for example,
calculate MAJOR.MINOR for ruby/php)
- simplify the code in webhook (get processed values without needing to
validate or compute anything)
- transaction - the values are all coming from the reconcile, so if the
runtime details are updated, there will be no race with the new result
of the reconcile and the webhook that also access these values.

This PR extract the functions to handle the distro params to a dedicated
file since the original function in the `sync` is quite long already.
It also introduces the "append env variables" found in CRI detection
into the params. They will later be used to refactor the pod manifest
env injection which currently uses otelsdk.

This is the error in UI when we have criError and env injection is set
to pod manifest:

<img width="738" height="218" alt="image"
src="https://github.com/user-attachments/assets/c3130a2f-30d0-4fc0-8c80-f55e9012aeb9"
/>