CORE-76: Migrate legacy Actions to generic Actions (#3456)
## Description

This PR updates all legacy Action controllers to simply be in charge of
migrating legacy Actions to the new Generic CRD.

If a legacy action is created or modified, a new Generic action will be
made for it with the same name prefiexed with `migrated-legacy-`. The
Generic action will have the legacy action as an ownerreference. No
legacy actions will be directly used in the generation of any Processor
configs, nor will they be listed as ownerreferences on the Processor
(instead the generated Generic action will take their place).

The UI migration to Generic Actions CRD in
https://github.com/odigos-io/odigos/pull/3450 means that no legacy
Actions will be visible in the UI. Because of that, there is no need to
continue supporting both code paths in each controller.

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [x] Added Unit Tests
- [ ] Updated e2e Tests
- [ ] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly

---------

Co-authored-by: Amir Blum <amirgiraffe@gmail.com>