perf: add k8s api cache with sources (#3466)
Fixes: CORE-250

This PR introduces a cache to the frontend component, so it can utilize
the cache client to pull in `Source` CRD fast and without bothering the
api server.

This will help with the following issues:

- faster responses (less roundtrip to api server for fetching resources)
- less calls -> less chance to hit rate limit (as seen in few clusters)
- lower footprint on api server (it still needs to setup a watcher and
update the cache on changes, but we will not hit a storm of requests on
api server when users interact with the UI)

There is always the issue with memory concern, as these objects are now
always in memory even when not used. Since this is our CRD, and it is
expected to be small in size and we will have at most one object per
workload/ns in the cluster, I am sure it is not a problem for the
memory.

Going forward, if we want to add more objects to the cache, we will need
to monitor the memory and make sure it is not crossing the limit set for
the container