ci: run links check only on release PRs only (#3515)
Fixes: CORE-262

Small suggestion I want to add.
Since the link checks workflow tends to fail from time to time because
of network and transient issues, and rarely catch real issues (I don't
remember even one to be honest :)). I want to suggest running it only on
release PRs.

If there is any issue, we will have the chance to catch that before
starting a release, but our day-to-day life will be less noisy. Open for
opinions on this change.