CORE-134: Add more options for GCP exporter + K8sConfiger interface (#3449)
## Description

This adds the following options to the GCP exporter:

* `timeout` -- configure timeout for requests to cloud observability API
* `project ID` -- used when sending telemetry to a different GCP project
than where the app is currently running (also conditionally required if
running off-GKE and sending traces to GCP observability)
* `billing project ID` -- the *billing* project for counting trace quota
usage when sending to GCP
* `Google application credentials` -- service account token/info for
sending traces to GCP when running off GKE

Application credentials requires something new where a Destination can
conditionally configure the Collector Deployment. This is because, when
setting custom credentials for GCP, the exporter needs the credentials
as a file in the Pod.

This is implemented with a new `K8sConfiger` interface that builds on
the common `Configer` to add k8s-specific updates to the deployment
without adding k8s dependencies to the common package. `K8sConfiger`
objects are linked to their common `Configer` with a `DestinationType()`
method.

Replaces https://github.com/odigos-io/odigos/pull/2904 with a more
simplified, decoupled approach to `K8sConfiger`.

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [x] Added Unit Tests
- [ ] Updated e2e Tests
- [ ] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly