[RUN-112] enhance odiglet un-privileged capabilities configuration (#3518)
Follow up to #3310 and #3331 this PR:
1. Makes the capabilities configuration more flexible using Helm values.
A default set of capabilities is configured in the values - these will
be used only in un-privileged mode. Users can add/drop capabilities
according to their environment if desired.
2. data-collection capabiliteis are added to support #3309 in
un-privilieged mode.
3. runtime-detector is bumped one version to reduce it's security
requirements - and it is now supports a fall back for cases where it
can't read the PID namespace.