handle data-collection config map creation in autoscaler reconciler, enhance config map provider in the collector (#3522)
This PR changes the node collector (data-collection) config map
creation:
* Following #3385 the config map was created by the helm/cli
installation. This caused some issures when no sources were present and
after `helm upgrade` the config map would get deleted. This PR move the
responsibility of creating the no-op config map to the autoscaler which
is already the component that is responsible for updating it.
* If no sources are present, the data collection controller now creates
the no-op config map.
* The autoscaler deployment is set to be the owner of the config map -
this guarantees that the config map will get deleted when odigos is
un-installed - before this, the owner was the collectorsgroup - but if
no sources are present there is no collectorsgroup object.

The cm config provider in the collector is enhanced by making it wait
for the first time it retrieves the config - until the config map is
present. The config map will now get created once the autoscaler starts
and the data-collection reconciler is triggered upon start-up and the
no-op config map is created. - This should not take a lot of time but in
case it is, the config provider will wait until the config map will get
created for the first time, and won't return an error resulting in the
collector entering CrashloopBackoff.

To improve indication of this possible initial waiting step a startup
probe is added to the data collection container - this start up probe
will not return a healthy status until the first config is loaded into
the collector.