CORE-227: Add GCP OTLP exporter (#3519)
## Description

GCP recently announced availability for their [Telemetry (OTLP)
API](https://cloud.google.com/blog/products/management-tools/opentelemetry-now-in-google-cloud-observability)
which supports the OTLP exporter instead of using the googlecloud API
exporter.

However, using the OTLP exporter with GCP also requires configuring the
googleclientauth extension and ensuring that the `gcp.project_id`
attribute is set on traces. When running Odigos off of GCP and sending
traces to the GCP OTLP endpoint, the same application credential auth is
needed as was added in #3449.

This adds support for this leveraging the same k8sconfiger interface
added in #3449

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [ ] Added Unit Tests
- [ ] Updated e2e Tests
- [x] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly