feat: add metrics settings section to collectors group with default values (#3553)
Fixes: CORE-271

This PR is a preparation to migrate the spanmetrics to node-collector.

Wether we inject spanmetrics right now is in each destination code. I
moved this setting to the destination manifest, and is using it during
reconcile to write on the node collectors group if metrics in general is
enabled, and if spanmetrics should be collected or not.

This will allow us to:
- consolidate "metadata" for destination into the manifest where it's
better documented
- more "generic" approach and less places to touch when changing
- easy to add new destinations to span metrics (just enable a boolean in
manifest)
- option for us to see the aggregated computation which will be used for
the collector generation
- manifest field can be used in UI / docs and other components and offer
better experience for users (futrue effort we can make)

I think the logic to compute the collectors group should be moved into
the autoscler, but that is out of scope for current PR and we can
consider doing down the road