chore: do spanmetrics in node collector instead of cluster collector (#3556)
Fixes: CORE-271

This is so that span metrics calculations are scoped only to the pods in
the same node. today we calculate time-series for the same metric and
pod in each instance of cluster collector.

This will help:
- making sure that each metric is generated only once
- over all do less work and use less memory since less metrics
computations in pipeline

The downside is that work is now done in the node collector instead of
the cluster collector (more sensitive to memory and cpu resource and
cannot scale).