Pods webhook improvments (#3533)
Fixes: CORE-270

Minor improvements for the pods webhook:

- Run only on "CREATE" events (no need to run on "UPDATE" since we do
our mutations only on created pods.
- Do not reinvoke the webhook - this was added to support more cases
when there are other web-hooks, but it is not tested, and I think does
not actually works at the moment. we can always re-introduce it after we
test and design it properly.
- Allow to set the `TimeoutSeconds` property for the webhook timeout.
default is 10 seconds and can be increased or decreased if any issues
are observed