chore: bump odiglet base image to use go1.25 (#3624)
Update the base image of odiglet to use go1.25.3.
In addition using the `trixie` variant of the image to update to the
latest stable version of Debian.
Updating CI to publish the base image also to GCP.
In a follow-up PR I plan to update the odiglet to use this new image -
and update all the CI and go.mod to use go1.25.