CORE-324 | chore: bump otel collector to v1.44/v0.138.0 (#3669)
Bump the collector dependencies to v1.44/v0.138.
The main motivation for this upgrade is to have the latest `pdata`
performance improvements around memory allocation and re-using allocated
objects.

* loki exporter is deprecated:
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/41413.
. https://grafana.com/docs/grafana-cloud/send-data/otlp/send-data-otlp/
The loki exporter is kept in the collector build to allow users to still
use these destinations - but we should follow up with changing the
destinations to use the modern OTLP variant.
* opencensus exporter is deprecated
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/42239
* routingprocessor is deprecated
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/42154

* Removed the untis from metrics name
https://github.com/open-telemetry/opentelemetry-go/issues/7039 - this
impacts the self metrics we report to the UI - the names are ajusted.