CORE-342: Support DeploymentConfigs for OpenShift (#3738)
## Description

In OpenShift, `DeploymentConfig` is a native type that predates native
k8s `Deployments`. They function very similarly and are essentially
obsolete since the introduction of `Deployments` in upstream k8s.
However, they are a popular object and have remained in use for years
despite deprecation in OpenShift 4.14 (worth noting that even though
they are technically deprecated, there is currently no timeline for
removal from OpenShift). This change adds support for DeploymentConfigs
on OpenShift.

Depends on https://github.com/odigos-io/ui-kit/pull/456 for full UI
support but most functionality is available here

For semantic conventions, this uses `k8s.deployment.name` for
DeploymentConfigs. This requires a new way to differentiate Deployments
from DeploymentConfigs in the throughput metrics, since previously it
was just based on the attribute key (so `deployment.name` looked for a
`Deployment` workload). The new attribute is `odigos.workload.kind` and
solves this purpose.

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [ ] Added Unit Tests
- [ ] Updated e2e Tests
- [x] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly