fix: bypass node collector metrics pipeline for span metrics (#3787)
Fixes: CORE-381

This is temporary fix, for spanmetrics to bypass the regular metrics
pipeline, and export directly to the gateway.

For span metrics, the host name and k8s cloud attributes are already
attached on the spans, so they do not need to be re-added.

A user might choose to "shape" it's span metrics dimensions, and exclude
specific attributes - in which case, we do not want to re-add them.

This is a temporary workaround. The final goal is for user to choose
which cloud attributes to attach to metrics/spans, and have unified
behavior for the entire metrics as a whole (instead of fiddling with the
span resource attributes on span metrics)

Co-authored-by: Tamir David <tamir@odigos.io>