fix: bypass node collector metrics pipeline for span metrics (#3787) (#3810)
Fixes: CORE-381

This is temporary fix, for spanmetrics to bypass the regular metrics
pipeline, and export directly to the gateway.

For span metrics, the host name and k8s cloud attributes are already
attached on the spans, so they do not need to be re-added.

A user might choose to "shape" it's span metrics dimensions, and exclude
specific attributes - in which case, we do not want to re-add them.

This is a temporary workaround. The final goal is for user to choose
which cloud attributes to attach to metrics/spans, and have unified
behavior for the entire metrics as a whole (instead of fiddling with the
span resource attributes on span metrics)

## Description

<!-- Short summary of the changes -->

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [ ] Added Unit Tests
- [ ] Updated e2e Tests
- [ ] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly

Co-authored-by: Amir Blum <amirgiraffe@gmail.com>
Co-authored-by: Tamir David <tamir@odigos.io>