CORE-408 | fix: add apparmor support for running odiglet in non privileged mode (#3845)
When running in non privileged mode, and the nodes have `apparmor`
enabled (which is usually the case for ubuntu nodes) - we need to
specify the `apparmor` profile to be `unconfined` for the odiglet. This
is required to perform `ptrace` for example:
```
audit: type=1400 audit(1763721375.668:55): apparmor="DENIED" operation="ptrace" profile="cri-containerd.apparmor.d" pid=15552 comm="odiglet" requested_mask="read" denied_mask="read" peer="unconfined"
```

Depending on the k8s version - the profile can be specified either as an
annotation on the `odiglet` pod template or (in. newer versions) in the
pod security context. ref:
https://kubernetes.io/docs/tutorials/security/apparmor/

The helm chart is also updated to allow users to config custom apparomor
profile values if they wish.

Nodes not having apparmor installed should not be affected: see
https://github.com/kubernetes/kubernetes/blob/master/pkg/security/apparmor/helpers.go#L30