feat: add option to specify otel distro name for specific containers of sources (#3837)
Fixes: CORE-403

This allow users to specify the otel distro name they want to use for
specific containers of specific source.

Can be used to diagnose or bypass issues of the default agent, and try a
different one that is more compatible.

to use, one should populate the source CR like this:

```
apiVersion: odigos.io/v1alpha1
kind: Source
metadata:
  name: frontend
  namespace: default
spec:
  containerOverrides:
  - containerName: frontend
    otelDistroName: java-community
  workload:
    kind: Deployment
    name: frontend
    namespace: default
```

support in UI should be added as a followup PR.