feat: add flag to force hostNetork to false for odiglet (#3859)
When running in k8s clusters with version lower than 1.26, the
`LocalTrafficPolicy` service feature is not availabe.
In these cases any of the ports used by the odiglet containers can cause
a conflict failure if one of them is taken in the host by another
process.
To help mitigate this, we add a flag to force `hostNetwork: false` in
the odiglet - even for k8s versions lower than 1.26, For newer versions,
this flag is a no-op.
Using this flag can cause some functionalities to be affected mainly the
OpAmp agents which rely on communicating with the local odiglet, or OSS
OTel agents that export to the local data collection.