fix: allow k8sattributes processor empty metadata list (#3861)
Fixes: CORE-435

This change makes it so if the k8sattributes processor config does not
contain `extract.metadata`, then it is populated with `k8s.pod.name`
which is a noop.

if the list is empty, the processor will use the default values for this
config option, which contains `k8s.deployment.name`. this is not what
intended and can pull in a lot of objects into the collector memory,
causing OOM.