refactor: own-metrics collection on node collector (#3867)
Fixes: CORE-438

Few improvements  to how node collector collects it's own metrics:

previous:
- scrape `/metrics` endpoints. that would make the metrics come in
prometheus format and lose some of their otlp properties and structure,
and require us to modify them in pipeline.
- scrape is done every 10 seconds and not configurable
- pod name added in processor
- no way for other components (odiglet) to send own metrics about itself

after:
- use "periodic" reader that export the data directly in otlp format to
localhost
- add a new endpoint for odigos own metrics, on port `44318` (similar to
the one for application telemetry `4318` for easy to remember). it is
used by the periodic reader, but also can be utilized later on to send
data from odiglet or other components.
- scrape interval is configurable via the helm value
`ownTelemetry.metrics.scrapeInterval` default is `10s`
- pod name and node name is added on the collector config yaml under
`service.telemetry.resource`, so it exists on the metrics by default,
and the `pod-name` processor is no longer used in the odigos-prometheus
pipeline (but kept as as for ui metrics).
- metrics verbosity increased to "detailed" to capture more info about
collector.