ci: do not build rhel on every pr (#3887)
Fixes: CORE-450

Currently we are building all components with both regular Dockerfile
and rhel dockerfile.

While this is very safe, it adds a lot of noise and cost more CI money. 

This change makes it so that we build the rhel only on release PRs via
the
https://github.com/odigos-io/odigos/blob/main/.github/workflows/full-build.yaml

and if something broke for red hat, we will still have the chance to be
aware and address it, and avoid starting a release where rhel has issues
building