RUN-120 | chore: bump simple demo images in tests to v0.1.32 and add go->dotnet context propagation assertion (#3890)
* bump simple demo images to use the latest version which includes a
call from membership to pricing.
* Update the context propagation queries in tests - to assert that edge
in the trace.
* Update the data streams test to move the membership service to stream1
and remove pricing from stream1. This is done since now membership is
not a leaf in the trace - and having the current setup cause errors in
the simple trace db - since the parent span of pricing (used to belong
to stream1) is membership (which used to not be in any stream) - hence
the parent span id was not present in the db, and the query failed.