CORE-451: refactor imagePullSecrets/odigos config for instrumentor and autoscaler (#3903)
Follow up refactor to https://github.com/odigos-io/odigos/pull/3894

This switches the Autoscaler and Instrumentor to get imagePullSecrets
directly from the Odigos Configuration rather than by static flags
passed at startup. This allows both of these controllers to react to
changes in the imagePullSecrets values.

It works by adding a reconciler for the effective-config to each
controller, and storing the config in a global variable. This eliminates
the need to do an extra API call, particularly for the Instrumentor's
pods webhook (which was calling the API to look up the current config on
every instrumented pod). Because of this, the Instrumentor can also use
this global variable for other settings that it was looking up in that
call too.