CORE-451: Pre-pull agents init container in Odiglet (#3913)
In https://github.com/odigos-io/odigos/pull/3894 and
https://github.com/odigos-io/odigos/pull/3903 we tried wiring
imagePullSecrets to the Agents injected with the k8s-init-container
mountmethod.

Ultimately this wouldn't work since ImagePullSecrets are namespaced,
unless the user copied their pull secrets to every namespace they want
to instrument.

This PR instead adds the agents image as a no-op init container in the
Odiglet. This will pre-pull the agents image onto every node so that
user apps don't need to have a pull secret since it is already cached on
the node.