CORE-00 | fix: explicitly set injected init container AllowPrivilegeEscalation to false (#3927)
When the mount method being used it `k8s-init-container`, our pods
mutating webhook will add the `agents` init container to the
instrumented pod.
This container may be subject to security policies that may enforce that
it is running in a non privileged mode and that
allowPrivilegedEscalation field is also set to false explicitly.
This PR explicitly sets this fields to false.