chore: Run prerelease checks on patch and relevant branches (#3920)
https://github.com/odigos-io/odigos/pull/3650 misconfigured the labels
check for the prerelease CI, so it was just comparing `[Array] =
prerelease`. This switches it to `contains()` to actually check for the
label

It also now runs the prerelease checks on patch releases. This updates
the very-dependencies and verify-release-blockers checks to only look at
the relevant release branch. So for example, if we are running a patch
release, we don't need to verify that there aren't any release blocks or
dependency bots on main in odigos-enterprise.