feat(url-templatization): add support for non templated wildcard and regex (#3932)
Fixes: CORE-460

url-templatization processor for the collector new feature (syntax for
templatization rules):

This will help in capturing a broder list of options in one rule,
instead of needing to repeat it again and again.

Example:
```
      - "/support-cases/{id}/disable"
      - "/support-cases/{id}/enable"
      - "support-cases/{id}/revoke"
      - "support-cases/{id}/update"
``` 

Now becomes:
```
      - "/support-cases/{id}/*"
```

The third segment will not be templated, but will match any value,
reducing manual chore of maintaining the values, and cpu cycles since
less rules needs to be evaluated.

User can also set regex for a section, for example for k8s api to limit
the resource group version:
```
/apis/*/regex:^v[0-9]+((alpha|beta)[0-9]+)?$/namespaces/{namespace-name}/*/{resource-name}
```