CORE-451: Remove image-pull-secrets flag from instrumentor helm template (#3946)
https://github.com/odigos-io/odigos/pull/3894 added `image-pull-secrets`
as a command line flag to the Instrumentor (similar to how it was done
in the Autoscaler). This was then refactored out for both Autoscaler and
Instrumentor in https://github.com/odigos-io/odigos/pull/3903.

In enterprise, it was added in
https://github.com/odigos-io/odigos-enterprise/pull/1753 and removed in
https://github.com/odigos-io/odigos-enterprise/pull/1769.

However, while the Helm chart changes in #3903 removed this flag from
[the
Autoscaler](https://github.com/odigos-io/odigos/pull/3903/changes#diff-0716b0814249b7f1dbfb2a36387f298b24f56d6e40876833d3e4f7adf81a3760),
it did not remove it from the Instrumentor, causing a crash on 1.12 for
installs using that flag.

This removes the flag from the Helm chart. It's worth noting that this
flag-based change was never merged to the 1.11 release branch because we
decided on this change before the final version got cherry picked
(https://github.com/odigos-io/odigos/pull/3914) which didn't use this
flag