CORE-444 | chore: refactor instrumentation manager Handler API and add ProcessGroup (#3923)
This PR includes:
* Refactoring the `Handler` interface used by the eBPF instrumentation
manager. Moving the `ConfigGroup` and Distribution resolvers to be
functions of the `ProcessDetails` interface.
* Add a notion of `ProcessGroup` in addition to `ConfigGroup`. A process
group can be used to instrument or un- instrument a batch of process all
relevant to the save group (k8s pod for example). This will be further
utilized once we add support for instrumenting and un-instrumenting
without a restart. In k8s, once an instrumentationConfig object is
deleted - we can infer the process group from the deleted IC name alone.
This makes it useful for un-instrumenting all the relevant processes to
the deleted instrumentation config.