Update prerelease PR checks to not depend on each other
In the prerelease PRs, we check for offsets, dependency-syncs, and release blockers. These should all only depend on the initial check job ("decide"), not each other.

When these depend on each other, and 1 fails, the others will be skipped. This can be annoying if the other checks fail too once we get them to run. So, run them all at once so we know all failures to save time.