feat: add headers collection config to agent container section (#3985)
Fixes: CORE-495

agent config is currently written mostly in instrumentation config under
SDKConfig, which is indexed by language.

there is also a section which is indexed by container name, which has
config section for traces. this PR also write the headers collection
config there, so we can consume it from the right place, and eventually
remove it from SDKConfig once all is updated