CORE-444 | no-restart instrumentation (#3915)
Add support for instrumentation without performing a rollout on the
workload.
* Adding an opt in for a distribution metadata that enables a
distribution to declare it does not require a rollout. Such
distributions do not rely on environment variables to be added to the
instrumented processes, and no mounts are required as well (for agent
files).
* An instrumentation flow for such distribution is: Source added ->
InstrumentationConfig created -> runtime detection -> agent enabled
controller -> InstrumentationConfig controller for sending
instrumentation requests to the instrumentation manager.
* The final step is different between "regular" distributions and those
that do not require rollout. It adds a `Request` channel that allows
sending instrumentation/un-instrumentation request to the manager's
event loop.
* For an instrumentation request, we first check the workload has a
distribution that support not-restart instrumentation. If that's the
case we scan for the relevant processes for the workload on the node
(multiple pods, containers are handled) - and send a request with all
the relevant PIDs. In such a case the instrumentation manager calls
`detector.TrackProcesses` which allows it to get exit events for such
processes and cleans them if the instrumented processes exit for any
reason.
* For un-instrumentation - we detect a deletion of
instrumentationConfig, and send the `ProcessGroup` that is required to
be un-instrumented (e.g Deployment/membership/default) - the
instrumentation manager has an internal mapping for all the active
instrumentations for that workload, and it cleans them.

Ideally we would like to have no dependency/requirement for the webhook
for distributions that support no-restart. The added complexity is when
the workload is instrumented and later in time, new pods of the same
workload come up - whether from scaling or user imitated a rollout - we
need to have a reliable way to identify new relevant processes and
instrument them. We still rely on the webhook for such cases - since it
simplifies the logic and allows us to easibly get notified for new
processes the start which are relevant for instrumentation. To fully
remove the dependency on the webhook for such distributions we would
need to add capabilites for the process detector to track relevant
containers (cgroups) - and notify it on relevant new containers when
they come up using a k8s informer.